jQuery version: add triggering of open/hide events
Adds support for triggering open and hide events on a selector so that
you can listen for when fancyzoom has opened and when it has closed and
perform any further logic or processing.

Events triggered for open and hide actions are `fancyzoom:open` and
`fancyzoom:hide`, respectively. These events are namespaced to fancyzoom
with a colon per jQuery convention.

By default, the open and hide events are fired on the `body` element,
but can be customized by passing in a selector to the `eventTarget`
option when intializing fancyzoom:

    $('a').fancyzoom({eventTarget: '#somediv'});